Move text highlight to utils
