+++ generatedTypes/src/utils/index.d.ts
@@ -1,0 +1,2 @@
+import * as TextUtils from './TextUtils';
+export { TextUtils };
