+++ src/.babelrc.json
@@ -9,6 +9,7 @@
         "alias": {
           "commons": "./src/commons",
           "helpers": "./src/helpers",
+          "utils": "./src/utils",
           "hooks": "./src/hooks",
           "optionalDeps": "./src/optionalDependencies",
           "services": "./src/services",
