+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+import * as TextUtils from './TextUtils';
+
+export {TextUtils};
