+++ tsconfig.json
@@ -23,6 +23,7 @@
       "react-native-ui-lib": ["index.d.ts"],
       "commons": ["./src/commons"],
       "helpers": ["./src/helpers"],
+      "utils": ["./src/utils"],
       "hooks": ["./src/hooks"],
       "optionalDeps": ["./src/optionalDependencies"],
       "services": ["./src/services"],
